Refactor: move test suite related classes to /regression/suite
This groups the test suite related classes with high cohesion into a common namespace
