+++ src/filter/filter-strategy.interface.ts
@@ -1,4 +1,4 @@
-import {TestCase} from '../regression/test-case';
+import {TestCase} from '../regression/suite/test-case';
 
 /**
  * Depending on what level of regression testing is aimed at, a strategy may filter certain entries of the sitemap.
