+++ src/regression/flow/test-case-evaluator.spec.ts
@@ -1,7 +1,7 @@
 import Mock = jest.Mock;
 import {TestCaseEvaluator} from './test-case-evaluator';
 import {AbstractHttpStack} from '../abstract-http-stack';
-import {TestCase} from '../test-case';
+import {TestCase} from '../suite/test-case';
 import {RegressionResult} from '../result/regression-result';
 import {HttpResponseStackBuilder} from '../http-response-stack-builder';
 import {RegressionResultStatus} from '../result/regression-result-status.enum';
