+++ src/regression/sitemap-regression-test.ts
@@ -3,8 +3,8 @@ import {Observable} from 'rxjs/Observable';
 import {RegressionResultSet} from './result/regression-result-set';
 import {RegressionResult} from './result/regression-result';
 import {ReporterStrategy} from '../reporter/reporter-strategy.interface';
-import {TestSuite} from './test-suite';
-import {TestCase} from './test-case';
+import {TestSuite} from './suite/test-suite';
+import {TestCase} from './suite/test-case';
 import {TestCaseProcessor} from './flow/test-case-processor';
 import {TestCaseEvaluator} from './flow/test-case-evaluator';
 import {HttpResponseStack} from './http-response-stack';
