+++ src/reporter/reporter-strategy.interface.ts
@@ -1,9 +1,9 @@
-import {TestCaseConfig} from '../regression/config/test-case-config';
+import {TestSuiteConfig} from 'test-suite-config.ts';
 import {RegressionResultSet} from '../regression/result/regression-result-set';
 
 export interface ReporterStrategy {
 
     setOptions(options: any): this;
 
-    report(testCase: TestCaseConfig, result: RegressionResultSet): Promise<void>;
+    report(testCase: TestSuiteConfig, result: RegressionResultSet): Promise<void>;
 }
